压缩插件Demo独立成一个头文件；
